+++ src/main/java/io/github/jamalam360/jamfabric/JamModClientInit.java
@@ -55,7 +55,7 @@ public class JamModClientInit implements ClientModInitializer {
             }
         }), ItemRegistry.JAM_JAR);
 
-        ModelPredicateProviderRegistry.register(ItemRegistry.JAM_JAR, new Identifier("jam_jar_full"), ((stack, world, entity, seed) -> Jam.fromNbt(stack.getSubNbt("Jam")).ingredientsSize() != 0 ? 1.0f : 0.0f));
+        ModelPredicateProviderRegistry.register(ItemRegistry.JAM_JAR, new Identifier("jam_jar_full"), ((stack, world, entity, seed) -> Jam.fromNbt(stack.getSubNbt("Jam")).getIngredients().size() != 0 ? 1.0f : 0.0f));
 
         NetworkingRegistry.init(true);
     }