+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - Added Jamtastic items to their own custom item group.
 - Use Machete to optimize JAR, cutting the final size from ~100kb to ~80kb.
+- Added particles and sounds to the jam pot.
+- You can now use any water bucket in the FAPI water buckets tag to fill the jam pot.
+- Items will no longer be decremented when in creative mode.
+- Large code refactor.
 
 ### Closed Issues